Demo app breakdown.
